requirements.txt
mysql-connector-python==8.0.33
Pillow==9.5.0
ttkthemes==3.2.2